OAK-10443: refactor AbstractDataStoreService so that it doesn't need the statisticsProvider OSGi reference (#1115)
