+++ oak-blob-cloud-azure/src/main/java/org/apache/jackrabbit/oak/blob/cloud/azure/blobstorage/AbstractAzureDataStoreService.java
@@ -19,10 +19,13 @@
 
 package org.apache.jackrabbit.oak.blob.cloud.azure.blobstorage;
 
+import org.apache.felix.scr.annotations.Component;
+import org.apache.felix.scr.annotations.Reference;
 import org.apache.jackrabbit.core.data.DataStore;
 import org.apache.jackrabbit.core.data.DataStoreException;
 import org.apache.jackrabbit.oak.plugins.blob.AbstractSharedCachingDataStore;
 import org.apache.jackrabbit.oak.plugins.blob.datastore.AbstractDataStoreService;
+import org.apache.jackrabbit.oak.stats.StatisticsProvider;
 import org.osgi.framework.Constants;
 import org.osgi.framework.ServiceRegistration;
 import org.osgi.service.component.ComponentContext;
@@ -32,10 +35,14 @@ import java.util.Hashtable;
 import java.util.Map;
 import java.util.Properties;
 
+@Component(componentAbstract = true)
 public abstract class AbstractAzureDataStoreService extends AbstractDataStoreService {
     private static final String DESCRIPTION = "oak.datastore.description";
 
     private ServiceRegistration delegateReg;
+
+    @Reference
+    private StatisticsProvider statisticsProvider;
 
     @Override
     protected DataStore createDataStore(ComponentContext context, Map<String, Object> config) {
@@ -63,6 +70,14 @@ public abstract class AbstractAzureDataStoreService extends AbstractDataStoreSer
             delegateReg.unregister();
         }
         super.deactivate();
+    }
+
+    protected StatisticsProvider getStatisticsProvider(){
+        return statisticsProvider;
+    }
+
+    protected void setStatisticsProvider(StatisticsProvider statisticsProvider) {
+        this.statisticsProvider = statisticsProvider;
     }
 
     @Override
