+++ oak-blob-cloud/src/main/java/org/apache/jackrabbit/oak/blob/cloud/s3/AbstractS3DataStoreService.java
@@ -25,10 +25,12 @@ import java.util.Map;
 import java.util.Properties;
 
 import org.apache.felix.scr.annotations.Component;
+import org.apache.felix.scr.annotations.Reference;
 import org.apache.jackrabbit.core.data.DataStore;
 import org.apache.jackrabbit.core.data.DataStoreException;
 import org.apache.jackrabbit.oak.plugins.blob.AbstractSharedCachingDataStore;
 import org.apache.jackrabbit.oak.plugins.blob.datastore.AbstractDataStoreService;
+import org.apache.jackrabbit.oak.stats.StatisticsProvider;
 import org.osgi.framework.Constants;
 import org.osgi.framework.ServiceRegistration;
 import org.osgi.service.component.ComponentContext;
@@ -38,6 +40,9 @@ public abstract class AbstractS3DataStoreService extends AbstractDataStoreServic
     private static final String DESCRIPTION = "oak.datastore.description";
 
     private ServiceRegistration delegateReg;
+
+    @Reference
+    private StatisticsProvider statisticsProvider;
 
     @Override
     protected DataStore createDataStore(ComponentContext context, Map<String, Object> config) {
@@ -65,6 +70,15 @@ public abstract class AbstractS3DataStoreService extends AbstractDataStoreServic
             delegateReg.unregister();
         }
         super.deactivate();
+    }
+
+    @Override
+    protected StatisticsProvider getStatisticsProvider(){
+        return statisticsProvider;
+    }
+
+    protected void setStatisticsProvider(StatisticsProvider statisticsProvider) {
+        this.statisticsProvider = statisticsProvider;
     }
 
     @Override
