+++ oak-blob-plugins/src/main/java/org/apache/jackrabbit/oak/plugins/blob/datastore/AbstractDataStoreService.java
@@ -25,13 +25,12 @@ import static org.apache.jackrabbit.oak.spi.whiteboard.WhiteboardUtils.registerM
 import java.util.Dictionary;
 import java.util.Hashtable;
 import java.util.Map;
+import java.util.Objects;
 
 import javax.jcr.RepositoryException;
 
 import org.apache.jackrabbit.guava.common.base.Strings;
 import org.apache.jackrabbit.guava.common.collect.Maps;
-import org.apache.felix.scr.annotations.Component;
-import org.apache.felix.scr.annotations.Reference;
 import org.apache.jackrabbit.core.data.DataStore;
 import org.apache.jackrabbit.core.data.DataStoreException;
 import org.apache.jackrabbit.oak.api.blob.BlobAccessProvider;
@@ -46,6 +45,7 @@ import org.apache.jackrabbit.oak.spi.whiteboard.CompositeRegistration;
 import org.apache.jackrabbit.oak.spi.whiteboard.Registration;
 import org.apache.jackrabbit.oak.spi.whiteboard.Whiteboard;
 import org.apache.jackrabbit.oak.stats.StatisticsProvider;
+import org.jetbrains.annotations.NotNull;
 import org.osgi.framework.BundleContext;
 import org.osgi.framework.Constants;
 import org.osgi.framework.ServiceRegistration;
@@ -53,7 +53,6 @@ import org.osgi.service.component.ComponentContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-@Component(componentAbstract = true)
 public abstract class AbstractDataStoreService {
     private static final String PROP_HOME = "repository.home";
     private static final String PATH = "path";
@@ -66,9 +65,6 @@ public abstract class AbstractDataStoreService {
     private Registration mbeanReg;
 
     private Logger log = LoggerFactory.getLogger(getClass());
-
-    @Reference
-    private StatisticsProvider statisticsProvider;
 
     private DataStoreBlobStore dataStore;
 
@@ -89,7 +85,8 @@ public abstract class AbstractDataStoreService {
         PropertiesUtil.populate(ds, config, false);
         ds.init(homeDir);
 
-        BlobStoreStats stats = new BlobStoreStats(getStatisticsProvider());
+        BlobStoreStats stats = new BlobStoreStats(
+                Objects.requireNonNull(getStatisticsProvider(), "statisticsProvider must be non-null"));
         this.dataStore = new DataStoreBlobStore(ds, encodeLengthInId, cacheSizeInMB);
         this.dataStore.setBlobStatsCollector(stats);
         PropertiesUtil.populate(dataStore, config, false);
@@ -124,16 +121,12 @@ public abstract class AbstractDataStoreService {
 
     protected abstract DataStore createDataStore(ComponentContext context, Map<String, Object> config);
 
-    protected StatisticsProvider getStatisticsProvider(){
-        return statisticsProvider;
-    }
+    @NotNull protected abstract StatisticsProvider getStatisticsProvider();
+
+    protected abstract void setStatisticsProvider(StatisticsProvider statisticsProvider);
 
     protected String[] getDescription(){
         return new String[] {"type=unknown"};
-    }
-
-    void setStatisticsProvider(StatisticsProvider statisticsProvider) {
-        this.statisticsProvider = statisticsProvider;
     }
 
     protected static String lookup(ComponentContext context, String property) {
